Refactor: Extract new mod `ops::resolve::resolution`
